+++ Models/SqlRendimentoConsignado/Cedentes.cs
@@ -10,7 +10,7 @@ namespace Servicing.Models.sql_rendimento_consignado
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int IdCedente { get; set; }
+        public long IdCedente { get; set; }
 
         [Required]
         [MaxLength(50)]
